S11P21D202-345 feat : 닉네임 불러오기 기능 구현

+++ backend/src/main/java/com/e22e/moya/user/repository/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends CrudRepository<Users, Long> {
-    Optional<Users> findById(int id);
+    Optional<Users> findById(long id);
     Optional<Users> findByEmail(String email);
 }
